requirements.txt
python-igraph==0.7.1.post6
SpiffWorkflow==0.4
